Use projected nonce for new txs
